store: Extract root module
